+++ WHATSNEW.md
@@ -1,6 +1,7 @@
 ### Added
 
-- Method inventory.get_item() to only get one single item from the inventory
+- Option `--smtp` for `fmg assign`
+- Method `inventory.get_item()` to only get one single item from the inventory
 
 ### Changed